+++ packages/common-lib/src/lib/Events.ts
@@ -24,20 +24,28 @@ import {
 	CustomAttendanceFieldEntryType,
 	CustomAttendanceFieldValue,
 	EventObject,
+	EventType,
 	MemberReference,
 	Permissions,
 	PointOfContactType,
-	RawEventObject,
+	RawResolvedEventObject,
 	RawTeamObject,
 	User,
 } from '../typings/types';
 import { Either, EitherObj } from './Either';
 import { Maybe, MaybeObj } from './Maybe';
-import { areMembersTheSame, hasOneDutyPosition, hasPermission, isRioux } from './Member';
-import { isPartOfTeam } from './Team';
+import {
+	areMembersTheSame,
+	hasDutyPositions,
+	hasOneDutyPosition,
+	hasPermission,
+	isCAPMember,
+	isRioux,
+} from './Member';
+import { isPartOfTeam, isTeamLeader } from './Team';
 import { complement, destroy, get } from './Util';
 
-export const isPOCOf = (member: MemberReference, event: RawEventObject) =>
+export const isPOCOf = (member: MemberReference, event: RawResolvedEventObject) =>
 	(member.type === 'CAPNHQMember' && (member.id === 546319 || member.id === 542488)) ||
 	areMembersTheSame(member)(event.author) ||
 	event.pointsOfContact
@@ -78,7 +86,7 @@ export const canSignUpForEvent = (event: EventObject) => (
 		Either.map(destroy),
 	);
 
-export const getURIComponent = (event: RawEventObject) =>
+export const getURIComponent = (event: RawResolvedEventObject) =>
 	`${event.id}-${event.name.toLocaleLowerCase().replace(/ /g, '-')}`;
 
 export const hasMember = (event: EventObject) => (member: MemberReference) =>
@@ -103,12 +111,12 @@ export const getAttendanceRecordForMember = (attendance: AttendanceRecord[]) => 
 
 export const canManageEvent = (
 	threshold: Permissions.ManageEvent = Permissions.ManageEvent.FULL,
-) => (member: User) => (event: RawEventObject) =>
+) => (member: User) => (event: RawResolvedEventObject) =>
 	isPOCOf(member, event) || effectiveManageEventPermissionForEvent(member)(event) >= threshold;
 
 export const canMaybeManageEvent = (
 	threshold: Permissions.ManageEvent = Permissions.ManageEvent.FULL,
-) => (member: MaybeObj<User>) => (event: RawEventObject) =>
+) => (member: MaybeObj<User>) => (event: RawResolvedEventObject) =>
 	Maybe.isSome(member) ? canManageEvent(threshold)(member.value)(event) : false;
 
 export const effectiveManageEventPermission = (member: User) =>
@@ -134,7 +142,9 @@ export const effectiveManageEventPermission = (member: User) =>
 		isRioux(member) ? Permissions.ManageEvent.FULL : Permissions.ManageEvent.NONE,
 	);
 
-export const effectiveManageEventPermissionForEvent = (member: User) => (event: RawEventObject) =>
+export const effectiveManageEventPermissionForEvent = (member: User) => (
+	event: RawResolvedEventObject,
+) =>
 	Math.max(
 		effectiveManageEventPermission(member),
 		isPOCOf(member, event) ? Permissions.ManageEvent.FULL : Permissions.ManageEvent.NONE,
@@ -162,3 +172,29 @@ export const applyCustomAttendanceFields = (eventFields: CustomAttendanceField[]
 			memberFields.find(({ title }) => title === customField.title) ??
 			defaultCustomAttendanceFieldValue(customField),
 	);
+
+/**
+ * See if someone has the ability to add attendeees to an event and modify attendance for
+ * an event
+ *
+ * @param member the member to check permissions for
+ * @param event the event that has attendance the member wants to modify
+ * @param team the team that belongs to the event, used for checking team leadership
+ */
+export const hasBasicAttendanceManagementPermission = (member: User) => (
+	event: RawResolvedEventObject,
+) => (team: MaybeObj<RawTeamObject>) =>
+	// Event management permissions
+	effectiveManageEventPermissionForEvent(member)(event) === Permissions.ManageEvent.FULL ||
+	// CAP duty positions
+	(isCAPMember(member) &&
+		hasDutyPositions([
+			'Cadet Executive Officer',
+			'Cadet Administrative Officer',
+			'Cadet Administrative NCO',
+		])(member)) ||
+	// Team leadership
+	(Maybe.isSome(team) &&
+		team.value.id === event.teamID &&
+		(event.type === EventType.REGULAR || event.targetAccountID === team.value.accountID) &&
+		isTeamLeader(member)(team.value));