+++ packages/common-lib/src/typings/apis/events/events.ts
@@ -31,7 +31,7 @@ import {
 	ExternalPointOfContact,
 	Member,
 	NewEventObject,
-	RawEventObject,
+	RawResolvedEventObject,
 } from '../../types';
 
 /**
@@ -47,7 +47,7 @@ export interface EventViewerAttendanceRecord {
  * Contains information needed to display an event in EVentViewer.tsx
  */
 export interface EventViewerData {
-	event: RawEventObject;
+	event: RawResolvedEventObject;
 	pointsOfContact: Array<DisplayInternalPointOfContact | ExternalPointOfContact>;
 	attendees: Array<APIEither<EventViewerAttendanceRecord>>;
 	sourceAccountName?: string | undefined;
@@ -57,6 +57,7 @@ export interface EventViewerData {
 		name: string;
 		accountName: string;
 	}>;
+	authorFullName: MaybeObj<string>;
 }
 
 /**
@@ -109,7 +110,7 @@ export interface Copy {
  * TODO: Allow this to be done
  */
 export interface GetList {
-	(params: {}, body: {}): APIEither<RawEventObject[]>;
+	(params: {}, body: {}): APIEither<RawResolvedEventObject[]>;
 
 	url: '/api/events';
 
@@ -166,7 +167,7 @@ export interface GetEventViewerData {
  * Get's the next meeting with the 'Recurring Meeting' activity type set
  */
 export interface GetNextRecurring {
-	(params: {}, body: {}): APIEither<MaybeObj<RawEventObject>>;
+	(params: {}, body: {}): APIEither<MaybeObj<RawResolvedEventObject>>;
 
 	url: '/api/events/recurring';
 
@@ -183,7 +184,9 @@ export interface GetNextRecurring {
  * Creates a link between the specified account and event
  */
 export interface Link {
-	(params: { eventid: string; targetaccount: string }, body: {}): APIEither<RawEventObject>;
+	(params: { eventid: string; targetaccount: string }, body: {}): APIEither<
+		RawResolvedEventObject
+	>;
 
 	url: '/api/events/:eventid/link/:targetaccount';
 
@@ -200,7 +203,7 @@ export interface Link {
  * Gets the next `X` events, where `X` is defined in the Registry as `Website.ShowUpcomingEventCount`
  */
 export interface GetUpcoming {
-	(params: {}, body: {}): APIEither<RawEventObject[]>;
+	(params: {}, body: {}): APIEither<RawResolvedEventObject[]>;
 
 	url: '/api/events/upcoming';
 
@@ -254,7 +257,7 @@ export interface Delete {
  * from Date.now()
  */
 export interface GetRange {
-	(params: { timestart: string; timeend: string }, body: {}): APIEither<RawEventObject[]>;
+	(params: { timestart: string; timeend: string }, body: {}): APIEither<RawResolvedEventObject[]>;
 
 	url: '/api/events/:timestart/:timeend';