+++ packages/common-lib/src/typings/types.ts
@@ -436,6 +436,11 @@ export enum GroupTarget {
 	ACCOUNT,
 }
 
+export enum EventType {
+	REGULAR = 1,
+	LINKED,
+}
+
 // tslint:disable-next-line: no-namespace
 export namespace Permissions {
 	export enum FlightAssign {
@@ -1008,7 +1013,7 @@ export interface DebriefItem extends NewDebriefItem {
 
 export type RawPointOfContact = InternalPointOfContact | ExternalPointOfContact;
 
-export interface RawEventObject extends AccountIdentifiable, NewEventObject {
+export interface RawRegularEventObject extends AccountIdentifiable, NewEventObject {
 	/**
 	 * ID of the Event, can be expressed as the event number
 	 */
@@ -1030,37 +1035,80 @@ export interface RawEventObject extends AccountIdentifiable, NewEventObject {
 	 * comment provided by a member
 	 */
 	debrief: DebriefItem[];
-	/**
-	 * If this is a linked event this will be present
-	 *
-	 * Linking events allows for one account to copy an event of another account
-	 * and receive updates and such
-	 */
-	sourceEvent?: null | {
-		/**
-		 * ID of the event it came from
-		 */
-		id: number;
-		/**
-		 * The account linked from
-		 */
-		accountID: string;
-	};
+
 	googleCalendarIds: {
 		/**
 		 * UUID of the Google Caldendar event on the main calendar
+		 *
+		 * Is null when the event is not to be published, such as a draft event
 		 */
 		mainId?: null | string;
 		/**
 		 * UUID of the Google Calendar registration deadline event on the main calendar
+		 *
+		 * Not present if there is no registration deadline
 		 */
 		regId?: null | string;
 		/**
 		 * UUID of the Goodle Calendar fee deadline event on the main calendar
+		 *
+		 * Not present if there is not a fee deadline
 		 */
 		feeId?: null | string;
 	};
+
+	type: EventType.REGULAR;
 }
+
+export interface RawLinkedEvent extends AccountIdentifiable {
+	/**
+	 * ID of the link to the event
+	 */
+	id: number;
+	/**
+	 * Who it is that established this link
+	 */
+	linkAuthor: MemberReference;
+	/**
+	 * The event ID that this link references
+	 */
+	targetEventID: number;
+	/**
+	 * The account ID that the event this link references belongs to
+	 */
+	targetAccountID: string;
+	/**
+	 * Used for quick rendering the calendar
+	 */
+	pickupDateTime: number;
+	/**
+	 * Used for quick rendering the calendar
+	 */
+	meetDateTime: number;
+	/**
+	 * Mirrors the information for a regular event object. This is present
+	 * because a Google calendar will need items for the linked events as
+	 * well
+	 */
+	googleCalendarIds: {
+		mainId?: null | string;
+		regId?: null | string;
+		feeId?: null | string;
+	};
+	/**
+	 * Specify that this is a linked event
+	 */
+	type: EventType.LINKED;
+}
+
+/**
+ * Represents the information stored in the database
+ */
+export type RawEventObject = RawLinkedEvent | RawRegularEventObject;
+
+export type RawResolvedEventObject =
+	| RawRegularEventObject
+	| (Omit<RawRegularEventObject, 'type'> & RawLinkedEvent);
 
 export type FullPointOfContact = DisplayInternalPointOfContact | ExternalPointOfContact;
 
@@ -1068,7 +1116,7 @@ export type FullPointOfContact = DisplayInternalPointOfContact | ExternalPointOf
  * The meat of what this website is designed for; events can be signed up for
  * and hold information to facilitate easy information distribution
  */
-export interface EventObject extends RawEventObject {
+export interface RegularEventObject extends RawRegularEventObject {
 	/**
 	 * New events start with no attendance, but there can be procedurally
 	 * generated attendance on the client side to include internal POCs
@@ -1079,6 +1127,8 @@ export interface EventObject extends RawEventObject {
 	 */
 	pointsOfContact: FullPointOfContact[];
 }
+
+export type EventObject = RegularEventObject | (Omit<RegularEventObject, 'type'> & RawLinkedEvent);
 
 /**
  * Used for transfer when creating a new event object, as it cannot know what
@@ -3962,22 +4012,22 @@ export interface StoredMFASecret {
 
 export interface SignInLogData {
 	/**
-	 * 
+	 *
 	 */
 	memberRef: MemberReference;
 
 	/**
-	 * 
+	 *
 	 */
 	accountID: string;
 
 	/**
-	 * 
+	 *
 	 */
 	lastAccessTime: number;
 
 	/**
-	 * 
+	 *
 	 */
 	accessCount: number;
 }