+++ packages/server/src/api/events/debrief/deletedebrief.ts
@@ -18,7 +18,7 @@
  */
 
 import { ServerAPIEndpoint } from 'auto-client-api';
-import { always, api, get, RawEventObject, SessionType } from 'common-lib';
+import { always, api, EventType, get, RawRegularEventObject, SessionType } from 'common-lib';
 import { getEvent, PAM, saveEventFunc } from 'server-common';
 
 export const func: (now?: () => number) => ServerAPIEndpoint<api.events.debrief.Delete> = (
@@ -26,15 +26,23 @@ export const func: (now?: () => number) => ServerAPIEndpoint<api.events.debrief.
 ) =>
 	PAM.RequireSessionType(SessionType.REGULAR)(req =>
 		getEvent(req.mysqlx)(req.account)(req.params.id)
-			.map<[RawEventObject, RawEventObject]>(oldEvent => [
-				oldEvent,
-				{
-					...oldEvent,
-					debrief: oldEvent.debrief.filter(
-						({ timeSubmitted }) => timeSubmitted !== parseInt(req.params.timestamp, 10),
-					),
-				},
-			])
+			.filter(event => event.type === EventType.REGULAR, {
+				type: 'OTHER',
+				code: 400,
+				message: 'You cannot modify the debrief items of a linked event',
+			})
+			.map<[RawRegularEventObject, RawRegularEventObject]>(
+				(oldEvent: RawRegularEventObject) => [
+					oldEvent,
+					{
+						...oldEvent,
+						debrief: oldEvent.debrief.filter(
+							({ timeSubmitted }) =>
+								timeSubmitted !== parseInt(req.params.timestamp, 10),
+						),
+					},
+				],
+			)
 			.flatMap(([oldEvent, newEvent]) =>
 				saveEventFunc(now)(req.configuration)(req.mysqlx)(req.account)(oldEvent)(
 					newEvent,