+++ packages/server/src/api/events/events/getevent.ts
@@ -18,12 +18,12 @@
  */
 
 import { ServerAPIEndpoint } from 'auto-client-api';
-import { api } from 'common-lib';
+import { api, Maybe } from 'common-lib';
 import { getEvent, getFullEventObject } from 'server-common';
 
 export const func: ServerAPIEndpoint<api.events.events.Get> = req =>
 	getEvent(req.mysqlx)(req.account)(req.params.id).flatMap(
-		getFullEventObject(req.mysqlx)(req.account)(req.member),
+		getFullEventObject(req.mysqlx)(req.account)(Maybe.some(req.account))(req.member),
 	);
 
 export default func;