+++ packages/server/src/api/events/events/linkevent.ts
@@ -19,12 +19,16 @@
 
 import { ServerAPIEndpoint } from 'auto-client-api';
 import {
+	AccountObject,
 	always,
 	api,
 	AsyncEither,
 	asyncRight,
 	errorGenerator,
+	EventType,
 	hasBasicEventPermissions,
+	RawRegularEventObject,
+	RawResolvedEventObject,
 	toReference,
 	User,
 } from 'common-lib';
@@ -34,28 +38,37 @@ export const func: ServerAPIEndpoint<api.events.events.Link> = req =>
 	AsyncEither.All([
 		getEvent(req.mysqlx)(req.account)(req.params.eventid),
 		getAccount(req.mysqlx)(req.params.targetaccount),
-	]).flatMap(([event, targetAccount]) =>
-		asyncRight(
-			PAM.getPermissionsForMemberInAccountDefault(
-				req.mysqlx,
-				toReference(req.member),
-				targetAccount,
-			),
-			errorGenerator('Could not get permissions for account'),
-		)
-			.map<User>(permissions => ({ ...req.member, permissions }))
-			.filter(hasBasicEventPermissions, {
-				type: 'OTHER',
-				code: 403,
-				message:
-					'Member does not have permission to perform this action in the specified account',
-			})
-			.map(always(targetAccount))
-			.flatMap(
-				linkEvent(req.configuration)(req.mysqlx)(req.account)(event)(
+	])
+		.filter(([event]) => event.type === EventType.REGULAR, {
+			type: 'OTHER',
+			code: 400,
+			message: 'You cannot link to a linked event',
+		})
+		.flatMap(([event, targetAccount]: [RawRegularEventObject, AccountObject]) =>
+			asyncRight(
+				PAM.getPermissionsForMemberInAccountDefault(
+					req.mysqlx,
 					toReference(req.member),
+					targetAccount,
 				),
-			),
-	);
+				errorGenerator('Could not get permissions for account'),
+			)
+				.map<User>(permissions => ({ ...req.member, permissions }))
+				.filter(hasBasicEventPermissions, {
+					type: 'OTHER',
+					code: 403,
+					message:
+						'Member does not have permission to perform this action in the specified account',
+				})
+				.map(always(targetAccount))
+				.flatMap(linkEvent(req.configuration)(req.mysqlx)(event)(toReference(req.member)))
+				.map(
+					linkedEvent =>
+						({
+							...event,
+							...linkedEvent,
+						} as RawResolvedEventObject),
+				),
+		);
 
 export default func;