+++ packages/common-lib/CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.16.0](http://github.com//cap-md089/evmplus-v6/compare/common-lib@1.7.0...common-lib@1.16.0) (2020-09-24)
+
+
+### Bug Fixes
+
+* fixed event website not displaying on the eventviewer ([9aeb97b](http://github.com//cap-md089/evmplus-v6/commit/9aeb97beff013dc4808e8c481155d95f1d7cd980)), closes [#43](http://github.com//cap-md089/evmplus-v6/issues/43)
+* fixed ImportCAPWATCHFile types to more accurately reflect the types of a CAPWATCH CSV file ([fffae38](http://github.com//cap-md089/evmplus-v6/commit/fffae38d6ba729c7592b1567242e4ea97cfe185e))
+* fixed linking events and attendance management ([78cb150](http://github.com//cap-md089/evmplus-v6/commit/78cb150039463240ef89b493d842d185c3b3d36a))
+* fixed time representation in Google calendar item description blocks ([c70773b](http://github.com//cap-md089/evmplus-v6/commit/c70773b5233aeba39d0656ae6d4eb87b423a5093)), closes [#44](http://github.com//cap-md089/evmplus-v6/issues/44)
+* removed Wing calendar ([b7f6d84](http://github.com//cap-md089/evmplus-v6/commit/b7f6d84f1a9f203297111d9d2386ee5569434894)), closes [#24](http://github.com//cap-md089/evmplus-v6/issues/24)
+* **sessions:** fixed session timeout not updating ([253e86d](http://github.com//cap-md089/evmplus-v6/commit/253e86d07422be45b9fb03a286ed067cece28eac))
+
+
+### Features
+
+* **signin:** added sign in logging ([c48fd90](http://github.com//cap-md089/evmplus-v6/commit/c48fd904a5b6c3f6a118ff74ebca88dc726dae41)), closes [#26](http://github.com//cap-md089/evmplus-v6/issues/26)
+* added attendancescanner ([70f9feb](http://github.com//cap-md089/evmplus-v6/commit/70f9feba454f823fb9d33a43e404eafa18fe64ab)), closes [#19](http://github.com//cap-md089/evmplus-v6/issues/19)
+* adding cadet promotion requirements display ([6659e57](http://github.com//cap-md089/evmplus-v6/commit/6659e571bcf938370ffb0954c053d24f907d24f4))
+* completed adding Cadet Promotion Requirements display on Main ([f2982b1](http://github.com//cap-md089/evmplus-v6/commit/f2982b1cc97dbce6d2b31f432e91499bfd39f627)), closes [#5](http://github.com//cap-md089/evmplus-v6/issues/5)
+* made PAM.RequireMemberType smarter ([fbbc6ed](http://github.com//cap-md089/evmplus-v6/commit/fbbc6eda975457e8fcee52c33d227c31e6ce5c7b))
+* provided better copy event dialogues ([00901ec](http://github.com//cap-md089/evmplus-v6/commit/00901ec4028c18a5d3cabc990439d493f322b14e)), closes [#13](http://github.com//cap-md089/evmplus-v6/issues/13)
+* **main.tsx:** continued work on adding cadet promotion requirements to main page ([47c96f3](http://github.com//cap-md089/evmplus-v6/commit/47c96f39f818e479c5ea774f21babec1322d9e6a))
+* added MFA using an OTPA ([900239b](http://github.com//cap-md089/evmplus-v6/commit/900239b673598e22194bd0cb0edbdfef6a3d4cfa))
+* **new:** continued development on cadet promotion requirements ([67c0a78](http://github.com//cap-md089/evmplus-v6/commit/67c0a7891e6fca07454c48a0a182054dd45e87e8))
+
+
+
+
+
 # [1.15.0](http://github.com//cap-md089/capunit-com-v6/compare/common-lib@1.7.0...common-lib@1.15.0) (2020-09-09)